`ubcli generateDDL`: prevent reading of database / roles list in case neither `-drop` nor `-create` is passed. This allows to init an existed DB for user without DBA privileges

+++ packages/ubcli/lib/initDB.js
@@ -138,23 +138,25 @@ function initDB (cfg) {
   const dbaConn = dbConnections.FAKE_DBA_CONN
 
   const generator = require(`./dbScripts/${dbDriverName}`)
-  let dbExists = generator.databaseExists(dbaConn, mainConnCfg)
-  if (cfg.dropDatabase) {
-    if (!dbExists) {
-      console.warn(`Database for connection ${mainConnCfg.name} not exists. Drop skipped`)
-    } else {
-      console.info(`Dropping a database for connection ${mainConnCfg.name}...`)
-      generator.dropDatabase(dbaConn, mainConnCfg)
-      dbExists = false
+  if (cfg.dropDatabase || cfg.createDatabase) { // read a databases / roles only for create/drop DB (can be allowed for DBA only)
+    let dbExists = generator.databaseExists(dbaConn, mainConnCfg)
+    if (cfg.dropDatabase) {
+      if (!dbExists) {
+        console.warn(`Database for connection ${mainConnCfg.name} not exists. Drop skipped`)
+      } else {
+        console.info(`Dropping a database for connection ${mainConnCfg.name}...`)
+        generator.dropDatabase(dbaConn, mainConnCfg)
+        dbExists = false
+      }
     }
-  }
-  if (cfg.createDatabase) {
-    if (!dbExists) {
-      console.info(`Creating a database ${mainConnCfg.name}...`)
-      generator.createDatabase(dbaConn, mainConnCfg)
-      dbaConn.commit()
-    } else {
-      console.warn(`Database for connection ${mainConnCfg.name} already exists. Creation skipped`)
+    if (cfg.createDatabase) {
+      if (!dbExists) {
+        console.info(`Creating a database ${mainConnCfg.name}...`)
+        generator.createDatabase(dbaConn, mainConnCfg)
+        dbaConn.commit()
+      } else {
+        console.warn(`Database for connection ${mainConnCfg.name} already exists. Creation skipped`)
+      }
     }
   }
